pyproject.toml
[project]
name = "golf-bros-app"
version = "0.1.0"
description = "This app was created to share you golfing experiences with like minded golfers. Think of it as golf facebook. enjoy!"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.1",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "pillow>=12.1.0",
]
